chore: trying to update web3-onboard deps
Hitting an issue when the latest version of @web3-onboad/common is installed, due to some version conflict (https://github.com/blocknative/web3-onboard/issues/1567).

Also hitting a problem with walletconnect where `global is not defined` when trying to connect a wallet. This problem was already happening on older versions (main branch), but trying to solve it at the same time.